feat: Moved most code over to new data structure (#23)
* feat: Moved most code over to new data structure, storing move in 2 bytes, 12bits for coordinates & 4 bits for promotions.
* feat: Speed increase storing pieces in bytes

+++ src/main/java/com/github/zerkath/rosemary/Evaluation/EvaluationCalculations.java
@@ -85,7 +85,7 @@ public class EvaluationCalculations {
 
   private int materialValue() {
     int result = 0;
-    for (Map.Entry<Integer, Integer> entry : state.pieceMap.entrySet()) {
+    for (Map.Entry<Byte, Integer> entry : state.pieceMap.entrySet()) {
       result += pieceToValue(entry.getKey()) * entry.getValue();
     }
     return result;
